Struct chacha20::ChaCha

pub struct ChaCha<R: Rounds, MC: MaxCounter> { /* fields omitted */ }

ChaCha family stream cipher, generic around a number of rounds.

Use the ChaCha8, ChaCha12, or ChaCha20 type aliases to select a specific number of rounds.

Generally ChaCha20 is preferred.