Struct chacha20::XChaCha

pub struct XChaCha<R: Rounds>(_);

XChaCha family stream cipher, generic around a number of rounds.

Use the XChaCha8, XChaCha12, or XChaCha20 type aliases to select a specific number of rounds.

Generally XChaCha20 is preferred.